Week2/requirements.txt
google-cloud==0.34.0
pandas==1.5.3
pandas-gbq==0.19.1
prefect==2.7.11
prefect-gcp==0.2.5
prefect-gcp[cloud_storage]==0.2.5
protobuf==4.21.12
pyarrow==11.0.0
SQLAlchemy==1.4.46
